moving ingest imagery to ingest.py
