+++ trunk/modules/python/dataset/reader.py
@@ -2,39 +2,6 @@ import os
 import numpy as np
 import theano as t
 
-def ingestImagery(filepath, shared=False, log=None) :
-    '''Load the dataset provided by the user.
-       filepath : This can be a cPickle, a path to the directory structure.
-       shared   : Load data into shared variables for training
-       log      : Logger for tracking the progress
-       return   :
-           Format -- 
-           (trainData, trainLabel), (testData, testLabel), labels
-
-           The 'trainLabel' and 'testLabel' are integer values corresponding
-           to the index into the 'labels' string vector. this provides a
-           better means to identify errors during back propagation, but
-           still allows label finding during classification.
-
-           TODO: Consider returning these as objects for more intuitive
-                 indexing. For now numpy indexing is sufficient.
-    '''
-    from dataset.shared import splitToShared
-    from dataset.pickle import readPickleZip
-    train = test = None
-
-    # Load the dataset to memory
-    train, test, labels = readPickleZip(filepath, log)
-
-    # load each into shared variables -- 
-    # this avoids having to copy the data to the GPU between each call
-    if shared is True :
-        if log is not None :
-            log.debug('Transfer the memory into shared variables')
-        return splitToShared(train), splitToShared(test), labels
-    else :
-        return train, test, labels
-
 def normalize(v) :
     '''Normalize a vector in a naive manner.'''
     minimum, maximum = np.amin(v), np.amax(v)
